third_party/WebKit/LayoutTests/fast/dom/promise-rejection-events-isolated-worlds.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';
async_test(function(t) {
  if (window.testRunner) {
    testRunner.evaluateScriptInIsolatedWorld(1, "var p = Promise.reject(new Error());");
  }
  window.onunhandledrejection = t.unreached_func('no events should fire');
  window.onrejectionhandled = t.unreached_func('no events should fire');
  setTimeout(function() { t.done(); }, 10);
}, 'promise rejection events should not be fired for content scripts');
</script>
